컨트롤러 구현

+++ dustin/src/main/java/io/dustin/apps/board/api/usecase/like/DeleteLikeUsecase.java
@@ -6,7 +6,9 @@ import io.dustin.apps.board.domain.like.service.ReadLIkeService;
 import io.dustin.apps.board.domain.like.service.WriteLikeService;
 import io.dustin.apps.common.code.BoardType;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 @Service
 @RequiredArgsConstructor
@@ -16,7 +18,10 @@ public class DeleteLikeUsecase {
     private final ReadLIkeService readLIkeService;
 
     public LikeDto execute(Long id, Long boardId, Long userId, BoardType boardType) {
-        Like like = readLIkeService.
+        Like like = readLIkeService.getLike(id);
+        if(!like.getUserId().equals(userId)) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "삭제 권한이 없습니다.");
+        }
         writeLikeService.delete(id);
         LikeDto dto = LikeDto.from(like);
         return LikeDto.from(like);
